feat: refactoring read_volumes to module kubernetes_services
