+++ app/v1/modules/kubernetes_services/__init__.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 
 
-from .script_read_env_from import ScriptEnvironReaderFromService
 from .script_to_dict import ScriptToDictParserService
+from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
+from .script_read_env_from import ScriptEnvironReaderFromService
 
 
 __all__ = [
-    "ScriptEnvironReaderFromService",
     "ScriptToDictParserService",
+    "ScriptVolumesReaderService",
+    "ScriptEnvironReaderFromService",
     "ScriptEnvironReaderService",
 ]
